# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talent Bridge API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/catalogs": {
            "post": {
                "tags": [
                    "Catalogs"
                ],
                "summary": "Create a new catalog item",
                "operationId": "93494050764076d0467ea0c27690027d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "catalog_name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "catalog_name": {
                                        "type": "string",
                                        "example": "Web Development Services"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Professional web development services with modern technologies"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catalog created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Catalog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "invalid create catalog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/catalogs/{id}": {
            "put": {
                "tags": [
                    "Catalogs"
                ],
                "summary": "Update a catalog item",
                "operationId": "5c40862116e524aa4644014666f869ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Catalog ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "catalog_name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "catalog_name": {
                                        "type": "string",
                                        "example": "Updated Web Development Services"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 600000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated professional web development services with modern technologies"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catalog updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes update"
                                        },
                                        "data": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catalog not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "catalog not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catalogs"
                ],
                "summary": "Delete a catalog item",
                "operationId": "7336d8617f47cba8040c2669b0a43d41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Catalog ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catalog deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Catalog deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catalog not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "catalog not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Catalog": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "catalog_name": {
                        "type": "string",
                        "example": "Web Development Services"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "description": {
                        "type": "string",
                        "example": "Professional web development services with modern technologies"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}